The string attributes can now be modified by the user.
The implementation is suboptimal; every time the user sets a string,
the current C string is freed and replaced by a copy of the
user-supplied string. When the user gets a string, a new Python object
is created from the C string (every time). But I dont want to make the
implementation more complex unless there's a real preformance need.